Move drunet to recon module
